ai-codebase-copilot/requirements.txt
gitpython
sentence-transformers
faiss-cpu
gradio
tree-sitter
tree-sitter-languages